fix reference cycle problem
