+++ backend/controllers/menu.controller.js
@@ -3,12 +3,10 @@ const Restaurant = require('../models/restaurant.model')
 const Menu = require('../models/menu.model')
 const Category = require('../models/category.model')
 const MenuItem = require('../models/menuItem.model')
-const {
-  present: presentMenuItem,
-} = require('../controllers/menuItem.controller')
+const { presentMenuItem } = require('../util')
 const _ = require('lodash')
 
-const present = obj => {
+const present = (obj) => {
   const { __v, ...data } = obj._doc
   return data
 }
@@ -24,7 +22,7 @@ readMenu = async (req, res, next) => {
 
     res.json({
       data: {
-        menuItems: menuItems.map(i => presentMenuItem(i)),
+        menuItems: menuItems.map((i) => presentMenuItem(i)),
         categories,
         ...present(obj),
       },
@@ -108,12 +106,8 @@ const findMenu = async (req, res) => {
 
 function validateUpdateDataFormat(menu) {
   const schema = {
-    name: Joi.string()
-      .min(1)
-      .max(50),
-    description: Joi.string()
-      .min(1)
-      .max(2047),
+    name: Joi.string().min(1).max(50),
+    description: Joi.string().min(1).max(2047),
   }
 
   return Joi.validate(menu, schema)
